requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
sqlalchemy
